Refactor analysis module: separate data types into analysis_types.py, enhance analysis_db with writer context manager, and streamline data handling in analyze_modules.py.
