+++ houdini/zcommon/src/zabob/common/__init__.py
@@ -2,8 +2,11 @@
 Common utilities for zabob houdini tools.
 '''
 
+from zabob.common.analysis_types import (
+    EntryType, HoudiniStaticData, ModuleData, AnalysisDBItem, AnalysisDBWriter
+)
 from zabob.common.analysis_db import (
-    analysis_db, get_stored_modules,
+    analysis_db, get_stored_modules, analysis_db_writer,
 )
 from zabob.common.common_paths import (
     ZABOB_COMMON_DIR,
@@ -44,7 +47,6 @@ from zabob.common.find_houdini import (
 )
 from zabob.common.infinite_mock import InfiniteMock
 from zabob.common.analyze_modules import (
-    EntryType, HoudiniStaticData, ModuleData,
     get_static_module_data, save_static_data_to_db,
     modules_in_path, import_or_warn,
 )
@@ -56,6 +58,14 @@ from zabob.common.detect_env import (
 )
 
 __all__ = (
+    "EntryType",
+    "HoudiniStaticData",
+    "ModuleData",
+    "analysis_db",
+    "get_stored_modules",
+    "AnalysisDBItem",
+    "AnalysisDBWriter",
+    "analysis_db_writer",
     "ZABOB_COMMON_DIR",
     "ZABOB_ZCOMMON_DIR",
     "ZABOB_HOUDINI_DIR",
@@ -109,7 +119,6 @@ __all__ = (
     "ModuleData",
     "HoudiniStaticData",
     "import_or_warn",
-    "get_stored_modules",
     "modules_in_path",
     "get_static_module_data",
     "save_static_data_to_db",
