+++ houdini/zcommon/src/zabob/common/analysis_db.py
@@ -2,11 +2,18 @@
 Definition and access for the analysis database.
 '''
 
+from abc import abstractmethod
 from contextlib import contextmanager
+from dataclasses import dataclass
 import sqlite3
 from pathlib import Path
 from collections.abc import Generator
 import sys
+
+from zabob.common.analysis_types import AnalysisDBItem, AnalysisDBWriter, HoudiniStaticData, ModuleData
+from zabob.common.common_utils import T, get_name, none_or
+from zabob.common.timer import timer
+
 
 @contextmanager
 def analysis_db(db_path: Path|None=None,
@@ -119,6 +126,120 @@ def analysis_db(db_path: Path|None=None,
             with sqlite3.connect(db_path) as conn:
                 init_db(conn)
                 yield conn
+
+
+@contextmanager
+def analysis_db_writer(db_path: Path|None=None,
+                       connection: sqlite3.Connection|None=None,
+                       ) -> Generator[AnalysisDBWriter, None, None]:
+    """
+    Context manager for writing to the analysis database.
+    If an existing connection is provided, it will use that connection,
+    and the caller is responsible for closing it.
+    If a database path is provided, it will create and initialize the
+    database at that path if it does not exist.
+    If neither is provided, it raises a `ValueError`.
+
+    Args:
+        db_path (Path): The path to the SQLite database file.
+        connection (sqlite3.Connection): An existing SQLite connection to use.
+    Yields:
+        AnalysisDBWriter: A callable that writes items to the analysis database.
+    """
+    # Insert or update static data
+    item_count = 0
+    cur_module: ModuleData|None = None
+    with analysis_db(db_path=db_path, connection=connection, write=True) as conn:
+        """
+        Context manager for writing to the analysis database.
+        Args:
+            conn (sqlite3.Connection): The SQLite connection to use.
+        Yields:
+            AnalysisDBWriter: A callable that writes items to the analysis database.
+            """
+
+        with timer('Storing') as progress:
+            cursor = conn.cursor()
+            item_count = 0
+            cur_module = None
+            def writer(datum: T, /) -> Generator[T|AnalysisDBItem, None, None]:
+                nonlocal item_count, cur_module
+                match datum:
+                    case ModuleData():
+                        if cur_module is not None:
+                            # Commit the previous module data
+                            cursor.execute('''
+                                UPDATE houdini_modules
+                                SET count = ?, status = ?, reason = ?
+                                WHERE name = ?
+                            ''', (
+                                item_count,
+                                'OK',
+                                None,
+                                cur_module.name))
+
+                        conn.commit()
+                        status = datum.status
+                        if isinstance(status, Exception):
+                            t = type(status)
+                            if hasattr(t, '__name__'):
+                                status = t.__name__
+                            else:
+                                status = str(t)
+                        cursor.execute('''
+                            INSERT OR REPLACE INTO houdini_modules (name, directory, file, count, status, reason)
+                            VALUES (?, ?, ?, ?, ?, ?)
+                        ''', (
+                                datum.name,
+                                none_or(datum.directory, str),
+                                none_or(datum.file, str),
+                                datum.count,
+                                status,
+                                datum.reason))
+                        item_count = 0
+                        if datum.status is not None:
+                            progress(f'Skipping module {datum.name} due to status: {datum.status}: {datum.reason}')
+                            yield datum
+                            return
+                        progress(f'Processing module {datum.name}...')
+                        yield datum
+                        cur_module = datum
+                        conn.commit
+                    case HoudiniStaticData():
+                        item_count += 1
+                        if item_count % 100 == 0:
+                            conn.commit()
+
+                        if datum.parent_name is not None:
+                            #print(f'Processing {datum.type} {datum.name} with parent {datum.parent_name}...')
+                            #if str(name) == 'BlendShape' or name == 'hou' or datum.type == EntryType.OBJECT:
+                            #    breakpoint()
+                            cursor.execute('''
+                                INSERT OR REPLACE INTO houdini_module_data (name, type, datatype, docstring,
+                                                                            parent_name, parent_type)
+                                VALUES (?, ?, ?, ?, ?, ?)
+                            ''', (
+                                datum.name,
+                                get_name(datum.type),
+                                get_name(datum.datatype),
+                                datum.docstring,
+                                datum.parent_name, get_name(datum.parent_type)))
+                        else:
+                            #print(f'Processing {datum.type} {datum.name} without parent...')
+                            cursor.execute('PRAGMA foreign_keys = OFF;')
+                            conn.commit()
+                            # Change line number
+                            cursor.execute('''
+                                INSERT OR REPLACE INTO houdini_module_data (name, type, datatype, docstring, parent_name, parent_type)
+                                VALUES (?, ?, ?, ?, NULL, NULL)
+                            ''', (datum.name,
+                                    get_name(datum.type),
+                                    get_name(datum.datatype),
+                                    datum.docstring))
+                            cursor.execute('PRAGMA foreign_keys = ON;')
+                            conn.commit()
+                        yield datum
+            yield writer
 
 
 def get_stored_modules(db_path: Path|None=None,
